+++ CHANGELOG.md
@@ -36,12 +36,15 @@ The main work (all changes without a GitHub username in brackets in the below li
     * Changed name of the unique storage id for servers or controllers added to MatterServer to "uniqueStorageKey"
     * Adjusted subscription callbacks to also provide the nodeId of the affected device reporting the changes to allow callbacks to be used generically when connecting to all nodes
     * Introduces a node state information callback to inform about the connection status but also when the node structure changed (for bridges) or such.
+  * Breaking: option "mdnsAnnounceInterface" was deprecated and replaced by "mdnsInterface" and now used to limit announcements and scanning to a specific interface
   * Feature: Enhanced CommissioningServer API and CommissioningController for improved practical usage
   * Feature: Makes Port for CommissioningServer optional and add automatic port handling in MatterServer
   * Feature: Allows removal of Controller or Server instances from Matter server, optionally with deleting the storage
   * Enhance: Makes passcode and discriminator for CommissioningServer optional and randomly generate them if not provided 
 * matter-node-shell.js
   * Feature: Completely refactored and enhances shell to support commissioning, identify and many more new commands. See Readme, try it
+* matter-node.js-examples
+  * Breaking: Rename parameter -announceinterface to -netinterface and use for announcements and scanning
 
 ## 0.6.0 (2023-10-08)
 * Matter-Core functionality: